intel_apl_adsp: make SOC_AGGREGATE_IRQ macro straight forward
Zephyr's multiple level interrupt encodings requires 2nd level
and up be off by 1 in the encodings. The SOC_AGGREGATE_IRQ macro
requires developers to increment the interupt number before
passing the number to macro. This is error prone so do
the increment in the macro itself, and developers can simply
pass in the actual interrupt number.

Signed-off-by: Daniel Leung <daniel.leung@intel.com>